update plotting scripts
